Get some game plan / data structures made for genome creation
For the neural net calculations, there's now a thorough game plan and a
set of data structures that'll make it hopefully easy to create the
genomes in a way that doesn't involve having to check retroactively to
ensure the genome is valid. It'll be guaranteed valid.

This will save time in each new random lifeform creation.
(Assuming copying a ~400 member hashset of (usize, usize) is faster than
 checking and redoing was on average)

But more importantly it'll save from evolving over lifeforms that were
never going to do anything anyways.