Update (again) the way neural net calculations work
Now it's much more recursive and much less dismissive of recursive
connections.

I've left a lot of the scratch work in this commit in case I want to
come back to it and make some modifications using some of the work that
was done during this push. This is basically an afternoon's worth of
work.