This commit is unstable
It is an attempt to break the UI into multiple screens, each rendered
next to each other. However what I didn't realize was that the UI
library here does not actually allow for multiple simultaneous screens
like widgets, each screen is the only one the engine can draw. So this
becomes invalid.

This is kind of the last straw for me with this UI library. I'm
switching now to tui (https://github.com/fdehau/tui-rs). This has orders
of magnitude more stars and activity and definitely can do multiple
simultaneous widgets.

I'm committing this work because it breaks a lot of the UI sections into
their own logical units, which I suspect will be helpful when I switch
libraries.