Alright, attacking is now a thing as well!
Also I saw an orgy and there became thousands of lifeforms. It only
happened on one run, but to prevent that, for now, I brought the number
of lfs created down to two. The gene pool clearly just turned into a
mate fest. Somehow there needs to be further checks put in by the
environment.

Next up is more asexual reproduction checking food.