Hey hey, I got pause better
I think this is the magic formula for pause to really work. Before the
CPU was still cooking a little bit. The sleep inside the loop was a hack
and made the interface lag. The CPU was still at 3%, which was better
than the 8% if there was no sleep. Now the CPU is at 0% when it's
paused.

Although technically it's not paused.. it'll still tick about once every
584 million years.