Make a number of UI changes:
It's not terribly pretty, in fact the code could definitely use to be a
lot nicer, but now the selected lifeform id is stable. So if you're at
the end of the list, as other lifeforms come and go, your selection
won't change. It's keyed by lf id, not index in the list of lfs.

Change visual symbols around

Clean up the neural_net drawing a little bit