Fix issue with tests occasionally failing on mutate
This happened before the previous commit, but working on that commit
alerted me to that issue. Every like 100 times the test was run it would
fail. This is because we were sometimes reassigning the same from or to
id to a gene, and the tests were expecting it to be strictly different.
And it should be strictly different. We choose whether to mutate or not
in another function. If we choose to, we should definitely mutate.