Fix the bug that was slowing things down
After a couple hundred thousand iterations, every time, the UI drawing
would slow to a crawl. The program ran just fine without the UI, but
even if the UI drawing was paused, or if the program ran a bunch first
before even starting the UI loop, the UI would crawl.

Turns out this was because I was never reducing the size of the events
vector. So the UI was trying to draw half a million event strings every
iteration.

Now it's like a ring buffer.